# docs/openapi.yaml
openapi: 3.0.3
info:
  title: HealthLink API
  description: AI-driven Comorbidity Management Platform API
  version: 1.0.0-mvp
  contact:
    name: HealthLink Team
    email: team@healthlink.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.healthlink.ai/v1
    description: Production server
paths:
...
  /icer/evaluate:
    post:
      summary: Evaluate ICER & INB
      tags: [ICER]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IcerEvaluateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IcerEvaluateResult'
components:
  schemas:
...
    IcerArm:
      type: object
      required: [cost, effect, effect_unit]
      properties:
        cost:
          type: number
          minimum: 0
          description: 成本（元）
        effect:
          type: number
          description: 效果值（QALY或DALY）
        effect_unit:
          type: string
          enum: [QALY, DALY, OTHER]
          description: 效果单位
    IcerThreshold:
      type: object
      required: [value, unit]
      properties:
        value:
          type: number
          minimum: 0
          description: 阈值数值
        unit:
          type: string
          enum: [CNY_per_QALY, CNY_per_DALY]
          description: 阈值单位
        source:
          type: string
          enum: [policy, literature, WTP_survey, config]
          default: config
          description: 阈值来源
    IcerDiscount:
      type: object
      properties:
        cost_rate:
          type: number
          minimum: 0
          maximum: 1
          default: 0.03
          description: 成本贴现率
        effect_rate:
          type: number
          minimum: 0
          maximum: 1
          default: 0.03
          description: 效果贴现率
    IcerUncertainty:
      type: object
      properties:
        se_cost_0:
          type: number
          minimum: 0
          nullable: true
          description: 对照组成本标准误
        se_cost_1:
          type: number
          minimum: 0
          nullable: true
          description: 干预组成本标准误
        se_eff_0:
          type: number
          minimum: 0
          nullable: true
          description: 对照组效果标准误
        se_eff_1:
          type: number
          minimum: 0
          nullable: true
          description: 干预组效果标准误
        corr:
          type: number
          minimum: -1
          maximum: 1
          nullable: true
          description: 成本效果相关系数
        samples:
          type: integer
          minimum: 0
          maximum: 10000
          default: 0
          description: 蒙特卡洛抽样次数
    IcerEvaluateRequest:
      type: object
      required: [comparator, intervention]
      properties:
        comparator:
          $ref: '#/components/schemas/IcerArm'
          description: 对照臂
        intervention:
          $ref: '#/components/schemas/IcerArm'
          description: 干预臂
        perspective:
          type: string
          enum: [societal, payer, provider]
          default: payer
          description: 分析视角
        threshold:
          $ref: '#/components/schemas/IcerThreshold'
          description: 自定义阈值
        discount:
          $ref: '#/components/schemas/IcerDiscount'
          description: 贴现率配置
        uncertainty:
          $ref: '#/components/schemas/IcerUncertainty'
          description: 不确定性分析参数
        equity_weights:
          type: object
          additionalProperties:
            type: number
          description: 公平性权重
    IcerEvaluateResult:
      type: object
      properties:
        icer_value:
          type: number
          nullable: true
          description: ICER值
        icer_unit:
          type: string
          nullable: true
          description: ICER单位
        dominance:
          type: string
          enum: [simple, extended, none]
          description: 支配关系
        decision:
          type: string
          enum: [accept, reject, inconclusive]
          description: 决策结果
        net_benefit:
          type: number
          nullable: true
          description: 净效益
        ceac_prob_accept:
          type: number
          nullable: true
          minimum: 0
          maximum: 1
          description: 成本效果可接受曲线概率
        policy_version:
          type: string
          nullable: true
          description: 使用的策略版本
        threshold_used:
          type: number
          nullable: true
          description: 使用的阈值
        assumptions:
          type: object
          description: 假设条件
        evaluated_at:
          type: string
          format: date-time
          description: 评估时间
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
...